# api/openapi.yaml
openapi: "3.0.3"
info:
  title: 快组校园 API v2
  description: |
    快组校园 RESTful API v2 - 大学生组队与人才匹配平台

    ## 核心规约
    - URL路径：全小写，单词间用中划线分隔，只使用名词复数
    - JSON字段：严格使用小驼峰(camelCase)
    - 鉴权方式：Header携带 `Authorization: Bearer {token}`
  version: 2.0.0
  contact:
    name: 快组校园团队
servers:
  - url: https://api.zhaoxiaokai.xyz/api/v2
    description: 生产环境
  - url: http://localhost:8080/api/v2
    description: 本地开发
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: 获取当前用户信息
      operationId: getCurrentUser
      responses:
        "200":
          description: 成功获取用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserVO"
    put:
      tags: [Users]
      summary: 更新个人资料
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDTO"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserVO"
  /users/me/certification:
    post:
      tags: [Users]
      summary: 提交学生认证
      operationId: submitCertification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentImgUrl]
              properties:
                studentImgUrl:
                  type: string
                  description: 学生证照片URL
                  example: "https://cdn.example.com/student/123.jpg"
      responses:
        "200":
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponse"
...
  /projects:
    get:
      tags: [Projects]
      summary: 项目列表/搜索
      operationId: listProjects
      security: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: keyword
          in: query
          description: 关键词搜索
          schema:
            type: string
        - name: schoolId
          in: query
          description: 学校ID筛选
          schema:
            type: integer
        - name: status
          in: query
          description: 项目状态筛选
          schema:
            $ref: "#/components/schemas/ProjectStatus"
        - name: direction
          in: query
          description: 项目方向筛选
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectPageResponse"
    post:
      tags: [Projects]
      summary: 发布新项目
      description: 创建后状态默认为"待审核"
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectDTO"
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectVO"
  /projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: 项目ID
    get:
      tags: [Projects]
      summary: 获取项目详情
      description: 返回包含浏览量、队长信息的数据
      operationId: getProject
      security: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectDetailVO"
    put:
      tags: [Projects]
      summary: 修改项目信息
      description: 仅队长可操作
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectDTO"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectVO"
    delete:
      tags: [Projects]
      summary: 删除/下架项目
      description: 逻辑删除
      operationId: deleteProject
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponse"
  /projects/{id}/applications:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: 项目ID
    get:
      tags: [ProjectApplications]
      summary: 查看某项目的申请列表
      description: 仅队长可见
      operationId: listProjectApplications
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ApplicationPageResponse"
    post:
      tags: [ProjectApplications]
      summary: 申请加入项目
      description: 触发订阅消息逻辑
      operationId: applyToProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: 自我介绍/申请理由
                  example: "我有2年的Java开发经验..."
                contact:
                  type: string
                  description: 联系方式
                  example: "微信: example123"
      responses:
        "200":
          description: 申请成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProjectApplicationVO"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码(从1开始)
      schema:
        type: integer
        default: 1
        minimum: 1
    SizeParam:
      name: size
      in: query
      description: 每页条数
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
  schemas:
...
    BaseResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 业务状态码，200成功，其他为错误码
          example: 200
        message:
          type: string
          description: 提示信息
          example: "操作成功"
    PageInfo:
      type: object
      properties:
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页条数
        total:
          type: integer
          format: int64
          description: 总记录数
        totalPages:
          type: integer
          description: 总页数

    # ==================== 枚举类型 ====================
    ProjectStatus:
      type: string
      enum: [PENDING, ACTIVE, REJECTED, CLOSED]
      description: |
        项目状态:
        - PENDING: 待审核
        - ACTIVE: 进行中/已上架
        - REJECTED: 已驳回
        - CLOSED: 已关闭
    ApplicationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
      description: |
        申请状态:
        - PENDING: 待审核
        - APPROVED: 已通过
        - REJECTED: 已拒绝
...
    AuthStatus:
      type: string
      enum: [UNVERIFIED, PENDING, VERIFIED, FAILED]
      description: |
        认证状态:
        - UNVERIFIED: 未认证
        - PENDING: 审核中
        - VERIFIED: 已认证
        - FAILED: 认证失败

    # ==================== 认证相关 ====================
...
    UserVO:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatarUrl:
          type: string
        phone:
          type: string
        email:
          type: string
        school:
          $ref: "#/components/schemas/SchoolVO"
        major:
          $ref: "#/components/schemas/MajorVO"
        grade:
          type: integer
        oliveBranchCount:
          type: integer
          description: 剩余橄榄枝数量
        authStatus:
          $ref: "#/components/schemas/AuthStatus"
        createdAt:
          type: string
          format: date-time
    UpdateUserDTO:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 50
        avatarUrl:
          type: string
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        email:
          type: string
          format: email
        schoolId:
          type: integer
        majorId:
          type: integer
        grade:
          type: integer

    # ==================== 项目相关 ====================
    ProjectVO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        schoolId:
          type: integer
        schoolName:
          type: string
        direction:
          type: integer
        memberCount:
          type: integer
        educationReq:
          type: integer
        isCrossSchool:
          type: boolean
        status:
          $ref: "#/components/schemas/ProjectStatus"
        createdAt:
          type: string
          format: date-time
    ProjectDetailVO:
      allOf:
        - $ref: "#/components/schemas/ProjectVO"
        - type: object
          properties:
            creator:
              $ref: "#/components/schemas/UserVO"
            viewCount:
              type: integer
              description: 浏览量
            applicationCount:
              type: integer
              description: 申请数量
    CreateProjectDTO:
      type: object
      required: [name, description, memberCount]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
        schoolId:
          type: integer
        direction:
          type: integer
        memberCount:
          type: integer
          minimum: 1
        educationReq:
          type: integer
        isCrossSchool:
          type: boolean
          default: false
    UpdateProjectDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
        direction:
          type: integer
        memberCount:
          type: integer
          minimum: 1
        educationReq:
          type: integer
        isCrossSchool:
          type: boolean
    ProjectPageResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/ProjectVO"
        pageInfo:
          $ref: "#/components/schemas/PageInfo"

    # ==================== 项目申请相关 ====================
    ProjectApplicationVO:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        applicant:
          $ref: "#/components/schemas/UserVO"
        reason:
          type: string
        contact:
          type: string
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        appliedAt:
          type: string
          format: date-time
    ApplicationPageResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/ProjectApplicationVO"
        pageInfo:
          $ref: "#/components/schemas/PageInfo"

    # ==================== 人才档案相关 ====================
...
    SchoolVO:
      type: object
      properties:
        id:
          type: integer
        schoolName:
          type: string
        schoolCode:
          type: string
    MajorVO:
      type: object
      properties:
        id:
          type: integer
        majorName:
          type: string
        classId:
          type: integer
...